config: remove unused cat help
every category and subcategory has title and help text, but the help
text for categories is actually unused since it is the help text of
its general subcat that gets shown in the page of options when you
select a category node in the preferences tree, and every category
has a general subcat.

(the qt preference tree construction code did fetch it, but it always
got overwritten by the text of the general subcat).

this removes the unused defines (except for two that are still used);
changes the entries in the table to empty strings (to avoid having to
change the helper functions for now); and tweaks the interface modules
to stop trying to fetch it.

for the two defines that have been kept, since these are now used
exclusively by general subcats, they have been renamed to include
'general' in their name for consistency. (yes, `SUBCAT_ADVANCED_MISC`
is a general subcat; that name inconsistency may be fixed later).

note that the new name `AGENERAL_HELP` is not perfectly consistent
with the rest in terms of the typical `INPUT_GENERAL_HELP` form, but
this is how much of the "advanced" and "playlist" defines are named.
it is consistent with the playlist item; it's not worth renaming all
of them to be perfectly identical, and this seemed best.

since the `SOUT_HELP` and `PLAYLIST_HELP` text was much more useful than
`SOUT_GENERAL_HELP` and `PGENERAL_HELP` respectively, and considering the
feedback in review, the latter defines have adopted the text of the former.