config: add new cats and subcats for use in remodelled core help output
To achieve the goal of remodelling the core help output upon the same cats
and subcats as the preference interface, we need to first introduce some
new cats/subcats for the option blocks that only apply to help text.

Note that empty strings rather than nulls are deliberately used in place
of no help text to avoid a translation null-deref.