chromecast: disable httpd timeout
Fixes #25151
Fixes #25192

This is a hotfix for VLC 3.0 only. A proper solution need to be found in
VLC 4.0 (likely using a hls or dash server).

From Martin:
If it works it ain't broken, I don't see why it should depend on a full
HLS or DASH server.

Presumably:
Fixes #25366
Fixes #24359
Fixes #25192