vout_window: add visibility reporting callbacks
Those callbacks are meant to be used by vout_window implementations in
environment where the rendering is constrained by the application state.

The design case for those events is iOS where application must not draw
onto its surface when they have been moved to the background, but it
should be generalizable to every cases where the window is in a state
such as no more rendering is needed for it, like in Wayland when the
window is not visible anymore.

To handle these other cases, the reporting function might need to also
report a callback and better handle state change to notify the callback
when the application is back to be visible and has drawn something. This
is currently not implemented for iOS, so not done.