VLCVideoUIView: handle application backgrounding
When the application is moved to the background, iOS require that no
OpenGL drawing commands are sent to the GPU. Previous ios.m
implementation and current VLCOpenGLES2VideoView have code to handle
this case, flushing the commandlist and disabling the OpenGL commands,
but it is unfit with the design of vlc_gl_t since vlc_gl_MakeCurrent
would only fail and drawing would still happen.

In addition, making vlc_gl_MakeCurrent fail because of this reason would
prevent using it when closing the display, which should never happen
since there are no other time when we can release the resources.

Instead, signal the core that we should not draw anymore with the
callbacks.

Co-authored-by: Maxime Chapelet <umxprime@videolabs.io>