video_output: conditionnally render on display
Don't prepare/display if rendering is disabled. Rendering can only be
disabled when taking the display_lock so the prepare/display cannot be
done partially because of this patch.

In addition, there's no point in trying to render a 0x0 sized picture
on the screen, which is either producing no results or can even produce
protocol errors in display servers so disable rendering in that case.