VLCOpenGLES2VideoView: handle UIView resize
The UIView resize was previously handled using mostly UIView framework.
This is now forwarded by the vout_window through the vout_window_Report*
mechanism, under display_lock, mimicking the behaviour of every other
window system integration.

Because no feedback is currently forwarded to the window when the vout
thread has drawn something, it means that we cannot synchronize the
rendering with the end of the resize event from the main thread, so the
resize is out-of-sync.

Having the UIView being resized by the resize callback is necessary
though to move the generation of buffers out of the MakeCurrent callback
and making it not faillible.

Note that the module can run both in the main thread and in the vout
thread. To avoid lock inversion between both, since everything involving
both can be done in the main thread, like the resize in this case, we
only need to always be sure to lock within the main thread when we
dispatch to the main thread.

This patch also removes the handling through system notification events
which were leading to deadlocks and were no necessary anymore since it
has been moved to the vout_window implementation.

Co-authored-by: Maxime Chapelet <umxprime@videolabs.io>