android: utils: load SurfaceTextureListener wrapper
The SurfaceTextureListener is currently meant to be implemented outside
of libvlccore since libvlccore doesn't ship java compiled code, so it is
loaded through the ClassLoader stored from JNI_OnLoad.

The wrapper implements what's required for the SurfaceTexture's
OnFrameAvailableListener interface and redirect it to C callbacks.

It will be used to implement the proper SurfaceTexture synchronization.