android: utils: setup Handler for SurfaceTextureListener
SurfaceTextureListener is calling the frame callback from a Handler
object, which is by default either using the looper from the current
thread, or the main looper. Since there is no looper installed in the
vout thread, the main looper is used. However, the main looper is also
used by the windowing code whose events require the display_lock, so the
display code cannot use this looper.

Instead, we create a thread within the decoder device to attach a new
looper. For the SurfaceTexture, this is only possible since API 21 in
Android so anterior version will not be able to use any looper or
SurfaceTextureListener. It is possible to workaround this limitation by
creation the SurfaceTexture within the looper thread, but it needs
additional java wrapping code to implement the Handler behaviour in the
looper thread, so is deferred to a later patchset.