android: utils: synchronize on updateTexImage
Create SurfaceTextureListener for new SurfaceTexture to synchronize
them in the opengl interop. It synchronizes the rendering of the picture
from MediaCodec with the call to UpdateTexImage(). This is not used for
Android devices with API < 21 for now.

Since the wait is blocking and decoder can kill the producer side, the
wait is taking into account the interruption state and will gracefully
fail if it has been interrupted because we'd need to kill the display.

Fixes #25361.