vout: call module prepare() after waiting
The vout thread loop consists in the following steps:
 - process control events
 - prepare the picture
 - wait until the target PTS
 - display the picture

The picture is prepared in advance because the preparation time may vary
a lot, which could cause the picture to be displayed at an inaccurate
time.

To avoid state changes (due to resize for example) between prepare() and
display(), the display_lock is held while it waits for the target PTS,
causing lock starvation. The lock could not just be released for
waiting, because otherwise the display state may change between
prepare() and display() (and the vout display API guarantees that
display() will always be called after prepare()).

However, the preparation is composed of several steps:
 - interactive filters (core)
       interactive filters are not impacted by display state changes
       (and do not lock the display_lock).
 - SPU rendering (core)
       SPU rendering duration may vary a lot (depending on whether there
       are SPU or not). However, it could be done without display_lock.
 - prepare() (module)
       In practice, the module prepare() duration does not vary
       significantly. Currently, it is responsible for texture upload
       (which can be time consuming), but we plan to move that part to a
       separate step.

Therefore, we can estimate the module prepare() duration and call it
just before display().

Also add a small wait after prepare() to call display() at the expected
time.

This will allow to release the display lock during waiting.