vout: release display_lock during waiting
Release the display_lock for rendering SPU and during the wait until the
call to prepare()/display().

When the display_lock is released, the display state may change (e.g.
its size). In that case, an "invalidated" flag is set so that the vout
thread interrupts its waiting, abandons the result of its work which
depends on the display state, and starts pre-rendering a new picture for
the new display state.

The flag is protected by the clock lock because the wait is done on the
clock condition variable (deadline may change on clock state changes).