vout: split RenderPicture()
The first part of RenderPicture() filters the input picture. This
process is independent of the display.

The second part renders SPU then calls prepare()/display() on the vout
display module. This depends on display properties (in particular its
dimensions and ratio).

In order to be able to "restart" the second part on resize, extract it
to a separate function.