iosvlc: add tap recognizer to play/stop
The tapRecognizer provides a way to play and stop the media player. More
specifically, it provides a way to check that the size of the window and
especially the size of the UIView from the opengl provider is correct
when stopping, disabling the window, resizing and playing the media
again.