player: wait for vout on Darwin systems
The video output code in VLC is locking in the following order:

 1/ During initialization

 - vout window
   - main thread lock at init
 - vout display
   - main thread lock at init

 2/ During runtime

 - main thread event
   - vout window
     - vout display

There is a deadlock that could happen while the vout display is
initializing and the vout window reports an event to the display.
This was tackled by the reportEvent handling function in the
VLCVideoUIView.m module.

But the locking order is different again in the user application.
Everything is asynchronous for the user, so locking in the video output
is not an issue most of the time, except during the release of media
players while the input is still running where the following lock order
could happen:

 User application:
 - main thread
 - libvlc_media_player
 - vlc_player
 - vlc_player destructor (joining here)

 vlc_player_destructor:
 - input thread / input resource
 - video output
 - vout window
 - main thread

So another lock inversion happens here. To avoid it, this patch change
the following for the user application side:

 User application:
 - main thread
 - libvlc_media_player
 - vlc_player
 - main thread loop restricted to VLC events (waiting for destructor
   thread to destroy the vout, leave the loop and almost return)
 - vlc_player destructor (joining here)

So window and display will be able to fullfill their initialization and
then be closed correctly.