video_output: remove conditional render_now value on redisplay
By default render_now is true, meaning we want to do the swap without any
delay. It should only be set to false when we switch to a new picture that is
not late, i.e. we should display it on time.

When using the "redisplay" mechanism either we want to redisplay in this call
and refresh is true, we don't need to set render_now to true, as it's the
default value. When we don't redisplay (refresh = false) we don't need to set
render_now because we are not going to do the rendering (it can't be the
first and we didn't drop the current frame in this call). We will return
VLC_EGENERIC.

So we can leave it to true regardless of what we'll do with the "redisplay".