video_output: never drop the first frame
PreparePicture() returns the first non-late frame or NULL. But if this is the
first frame and there's no frame after, we still want to display it.

This was somehow ensured by the force flag set on the first picture by the
decoder, via the b_force flag, but it's cleaner and more explicit here.