avcodec: don't mark the first picture as b_force
The internal b_first_frame is used to set the forced flag for the first ever
picture pushed (regardless of flushes) and the End Of Stream pictures.

Setting the flag on the first picture is useless since the vout will know if
it's the first after a flush or not and act accordingly.

It's not up to the decoder to chose the display policies. The vout already
forces the first not to be dropped and be displayed ASAP.