decoder: don't flush the video output before pushing a b_still picture
b_still is set on End Of Stream pictures. There's no reason to discard "future"
frames which should not exist in the first place.

There's even less reason to reset the clock of flush filters. If the last frame
is interlaced it needs the previous frame(s) to be displayed correctly.