lua: discard basedir and DIR_SEP on UWP
On UWP, the lua folder is deployed right next to the executable in the
app installation folder, but the path returned by config_GetUserDir
(which is basedir) points to the local packages folder, located in the
current user AppData folder. Therefore, discard basedir and the
directory separator, to allow VLC to find the lua folder.