include: vlc_vout_window: add new KMS window type
The new KMS window represents an output from the DRM subsystem into a
given connector through the specified CRTC, or in simpler term, an
output screen. It allows running a "window" even in the case where no
window manager are presents, but it also makes sense in windowed
environment thanks to DRM leasing features.