kms: abort early if drm is not available
This is not using drmAvailable() which only checks against minor=0,
since card0 might not be available or not support DRM while card1 might
support it and be selected by --kms-device.

In addition, print the debug information about the current underlying
DRM driver implementation used. Note that according to NVidia's DRM
documentation[^1]:

    DRM-NVDC does not populate the drmVersionPtr structure's date,
    major, minor, and patchlevel fields.

[^1]: https://docs.nvidia.com/jetson/l4t-graphics/group__direct__rendering__manager.html#gaabb52f28bfc81b9af64866f88131d513