test: vlc-window: add test driver for vout window
The test driver can be used to check against the screen output list or
to check whether the window module can be opened correctly, allowing
testing for real-world environment module implementation.

Since it's also a separate binary, it can be used in development cycle
to check the behaviour of new vout window module implementation.