ogg: fix use of invalid pointers upon `asprintf()` failure
If `asprintf()` fails, the pointer is left in an unknown state. Should it
fail when trying to construct the content type and/or role strings, the
code would end up misusing the invalid pointers when trying to write data
into the buffer to build the object, and when freeing the strings at the
end.

The preference given in review was for the function to take a failure path
should these allocations fail. Note that the one caller already detects
failure via a null object/buffer pointer being returned, as with the
`calloc()` failure path, so this pointer is set to null prior to the new
returns.