subtitles: make filter function more logical
now returns `true`/`false` as to whether or not to filter the given file
as not being recognised as a subtitle file.

the output logic has thus now flipped.

i also took the opportunity to clean it up a little with better var names
and such since it's so tiny.