subtitles: restore missing check to path list splitting
this reverts commit 3743491c607b583b055ea44590e07e2b14171514, though
restoring it to a slightly different location - within the preceeding
condition block - since it is only necessary under that condition.

it turns out that i wasn't thinking hard enough; this is actually
necessary to catch strings that start with a comma or contain a
sequence of commas, otherwise you end up with one or more empty string
entries in the array.

i overlooked the fact that actually the condition can be true in the case
that `strchr()` finds the comma in the very first character of the string,
and thus replaces that first character with the null, thus turning the
current item (`psz_subdir`) into an empty string.