subtitles: handle path directory separator termination once per directory
instead of once per subtitle file, per directory.

this moves the handling to the function that splits the list of 'extra'
paths listed in the config settings. the other directory checked for
subtitles files - the one that the video file is located within - is
already guaranteed to end with a directory separator so we have no need
to worry about that one in moving this.