subtitles: cleanup var names
trying to understand precisely what is going on in this function is made
much harder by very messy and confusing variable names, particularly in
the first half.

much of this code was touched in the previous few commits anyway, so i
i thought i might as well tidy it up while i'm here.

i also covered a few addditional variables in the latter half because why
not be thorough.