subtitles: cleanup path splitter function
much of this function was rewritten in the previous commit and subsequent
ones, and the param var names are extremely uninituitive, so might as well
take the opportunity to clean them all up.

note, opportunity was taken to drop `k` since we can just use `i` if we
reset it afterwards.