qt: untangle hotkey editor from `ConfigControl`
The only reason that the hotkey editor derived from `ConfigControl` was due
to it being dynamically placed within the advanced preferences interface in
exactly the same way as controls for individual options. Now that this is
no longer the case, we can untangle it.

Doing so helps support later work.

Note that setting uniform row height had to be disabled, otherwise the
table appeared empty. I've encountered this issue before with other
preference interface tables. I think it's a Qt bug.