audio output: assume PCM output doesn't handle emphasis
The S/PDIF and HDMI output may carry of the emphasis type for MPEG Audio and
maybe other codecs. For PCM it may be possible with proper driver support or by
setting the HDMI header bit properly.

Since we try not to touch the codec inside non-PCM output, we don't want to
force a de-emphasis filter anyway.