adaptive: use std::unique_ptr<>
Represent the Segment and SubSegment ownership through
std::unique_ptr<>, using the existing std::move() to ensure it is
correctly forwarded and not leaked.

It makes it obvious when the Segment or SubSegment must be stored and
when we only have a transient reference to one. It also makes most of
the memory handling inexistant through RAII.