test: thread: set default asan options to use_sigaltstack=0
use_sigaltstack=1 has been the default in ASAN for a few year, but
starting with glibc 2.34, cancellation (because of setjmp/longjmp usage)
is corrupting the stack allocated by ASAN, leading to check failure when
ASAN is ready to check the application state. It does happen both with
VLC itself, and the tests doing cancellation, making test_thread fail in
particular.

Disable sigaltstack option in ASAN by default to avoid issue with thread
cancellation corrupting the ASAN stack markers.