aout: add gapless support
Add the ability to create a stream in gapless mode (gapless = true from
vlc_aout_stream_New()).

In case of gapless mode, this function will:
 - return NULL if the format does not match
 - return a stream in gapless state if the format match

A stream in gapless statte should notify the main stream that it is
ready to continue the playback by calling vlc_aout_stream_Play() a first
time. This function will set the gapless stream ready and return -EAGAIN
(without consuming the data). The user (DecoderThread) will then wait
for the gapless transition.

When the main stream (the playing one) is draining, it will first check
if a stream in gapless state is ready (it called
vlc_aout_stream_Play()). If there is no gapless stream, it will drain as
usual. If there is a gapless stream, it won't do anything and prepare
for the gapless transition.

Refs #549