input: add gapless support
When started with the INPUT_TYPE_GAPLESS type, an input will start its
buffering and wait for the input_WakeGapless() call. Decoders started in
gapless mode will wait for the end of the gapless state before outputing their
first outputs. The audio decoder started in gapless mode will setup the audio
output used by the current input for gapless and wait for the transition
just before playing the first buffer.

Note: If there is not valid audio data during the input buffering, the audio
gapless won't be possible.