nvdec: only load the CUDA DLL/functions once and keep them
cuInit is leaking and there's no way to release its memory. So we keep that
initialized CUDA setup in memory to avoid creating many other ones. It will
still be a single leak, but bigger.

Unloading the DLL, reloading the DLL and not calling cuInit for it might not
produce the desired effect. Given it's allocating memory it's probably using it
for something that is not there anymore the second time.

We load everything once and keep the functions in a static structure.

No need for opaque pointers relying on p_sys set early anymore.