opengl: interop_vaapi: link EGL directly
Use EGL symbols from EGL library directly instead of loading through
vlc_gl_GetProcAddress. In the future, we might want to use runtime
loading through the egl provider capabilities but it's currently unused
and loading EGL through GLX (vlc_gl_GetProcAddress) is leading to
undefined results and crash.

With this patch, the interop_vaapi correctly fails when running with GLX
and it falls back on EGL to load the interop.

Fixes #26813