video_output: drop late pictures coming from the static filters
If a deinterlacer takes too long we should also drop the pictures. Hopefully we
will start the deinterlacing sooner the next time (thanks to
chrono.static_filter).

Unlike pictures dropped out of the decoder, we don't need to flush static
filters since we passed them all the input pictures in order.

It's dropped like a late frame to display, not late from coming out of the
decoder.