vlc_stream: don't loop in vlc_stream_Read() on non retry errors
Only EINTR and EAGAIN errno are supposed to try again to do the same call.

For EINTR we already check if vlc_killed() is set. Otherwise it's another
interruption we don't care about and we should retry.

Other errno values should no retry to do the same calls. They probably produce
the same error on the next call and we'd loop on errors.