vlc_network: add helper to set network errors
Some internal calls return -1 and set an error as if they were socket APIs or
mixing their error codes with error codes from the socket API calls.

The caller doesn't know if the error is in errno or the network errno which is
different on Windows. So we should always read from the network errno when
there's a possible mixup.

net_errno represents the network errno but we can't write it with the existing
define on Windows.