vlc_network: add helper to check network errors
Some internal calls return -1 and set an errno as if they were socket APIs or
mixing their error codes with error codes from the socket API calls.

The caller doesn't know if the error is in errno or the network errno which is
different on Windows. So we should always read from the network errno when
there's a possible mixup.

net_errno represents the network error but if the value is set from a Windows
call, the test assuming it's a POSIX error will not read it properly. So when
checking POSIX error values, we translate the Windows ones to POSIX values.