udp: use vlc_net_set_errno() to set custom network errors
These calls are mixed with system socket calls, so we should put the errors
in the same place so the caller doesn't have to wonder which to use.

* net_SetMcastOut is called from net_ConnectDgram() which behaves like a
POSIX network call. The error value is not checked though.

* net_SourceSubscribe is called from net_OpenDgram() which reports the
net_errno value from the net_SourceSubscribe call.

* net_Subscribe is called from net_ListenSingle() and net_SourceSubscribe().
The caller of net_ListenSingle (via net_OpenDgram) doesn't check the net_errno.
net_SourceSubscribe needs to set the net_errno value (see above)

* net_SetDSCP is called from net_ConnectDgram() which behaves like a
POSIX network call. The error value is not checked though. It should behave
like getsockname() and setsockopt() used in the function, so set the net_errno.