access: dvb: split the internal loop error checks
On vlc_killed() it's equivalent to the poll returning -1 with errno at EINTR.
In that case we return VLC_EGENERIC.

scan_IsCancelled() failing is equivalent to a timeout as the poll will never
return any available data anymore.

We don't need to set a default value for i_ret anymore as we return whenever
there's an error.