contrib: vulkan-loader/header: update to 1.3.227
It contains the proper DLL name in mingw

The msys2 transforms everything into msys2 style DLLs but we should use the
system vulkan-1.dll found in windows/system32.

We should also not patch the .def file to make it use the wrong DLL name.

The ASM handling was fixed in the previous patch.