contrib: protoc: build without the host name
protoc generates platform-independent source code. We don't need to have one
version per target OS.

As explained in 104ac3901c28c2ef195047984dc7c8aa834e8ea5, we still need this
over just picking the right tool with AC_CHECK_TOOL(). Otherwise the PROTOC
variable doesn't have the PATH in it and we would need to keep the environment
PATH the same as it call to 'make'.

If protoc doesn't come from the contribs, we assume it's always in the PATH.