contrib: rav1e-vendor: build locally if not downloaded
If the tarball is not there, it will be built in a temporary directory and
put the the tarballs folder for future use.

We don't generate/keep the sha512sum since we build it locally, we just trust
it. Once that tarball is uploaded to the videolan contribs we can add a hash
to verify that download.

Rebuilding depends on .sum-rav1e so that we are using only a rav1e package with
the proper checksum.

Once we have the tarball we extract it so it has a rav1e-vendor/vendor
directory hierarchy.