vlc_serializer: helper to serialize and deserialize data types
Introduce a serializer helper which is used with the module of
serialization.
This permit to serialize and deserialize types implemented in
modules/misc/serializers/.

In case of serialization, type_serializer_t embedded type, will be
serialized in the memstream destination.

In case of deserialization, embedded type will be initialized from the
serialized data provided by type_buffer, buffer_size should match the
type_buffer length