mediainfo: introduce preparse process
The preparse process handle preparse requests, it read
infinitely on stdin.

Request format follows:
    request protocol:
    1. request size   : unsigned int
    2. mrl length     : unsigned int
    3. mrl            : char[]
    4. preparse depth : int

When a request is received, it use libvlc and create a media based on
the mrl received.

It define callbacks on preparse events to serialize the resulting
objects of the preparse phase.

Answers vary according which parsing event are triggered.

Type of answers:

 - subitem_tree_added:
    1. event type                 : libvlc_event_e
    2. number of subitems         : unsigned int
    3. size of serialized subitem : unsigned int
    4. serialized subitem         : char[]

 - media_parse_changed:
    1. event type                 : libvlc_event_e
    2. number of subitems         : unsigned int
    3. size of serialized subitem : unsigned int
    4. serialized subitem         : char[]
    ...

Answers are written on stdout.