mediainfo: handle vlc events dynamically
When the medialibary module preparse a media, it subscribe to
vlc_InputItemAttachmentsFound, to extract the embeddedthumnails.
The playqueue in vlc, don't suscribe to this event.

Both will use the preparse process, so preparse process has to be
configurable regarding which events it treat.

Events subitemtree_added and parse_changed are always handled, so
they are not included in request protocol.

request protocol:
1. request size     : unsigned int
2. number of events : unsigned int
3. array of events  : vlc_event_type_t[]
4. mrl length       : unsigend int
5. mrl              : char[]
6. preparse depth   : int

This commit implement the thumbnail found event:
1. event type                    : libvlc_event_e
2. number of attachments         : unsigned int
3. size of serialized attachment : unsigned int
4. serialized attachment         : char[]

The goal was to make the implementation of events handler scalable
easily.