vout: add vout_PreviousPicture
Simple approach: check all decoded frames until we reach a frame that
have the same date as the displayed one. Then, display the frame that
was just before.

Send a status to the decoder in case of success/failure since the input
need to know if he needs to seek again.