input: add input_PreviousFrame
Simple and costly approach (IO, CPU): Seek backward and tell the vout to
search for the frame that is just before the displayed one.

Listen to vout events via the es_out to adjust seek requests.

Ideally, just one seek request is needed but this is not always the case
since it depends on initial position of the demux/access or if the
picture if far from the gop...

The number of seek requests and the time needed is logged in debug.