tools: bootstrap: fix version checking algorithm
If major number is not equal for needmajor and gotmajor, the version
checker returns old, which can lead to faulty results.

For example:
check_version meson 1.0.0 0.6.0 returns old; where expected is ok.

Fix this issue by changing the check of ne to gt. This will revert
commit: ee5fad6a So that it returns old if needmajor is greater than
gotmajor.