contrib: prefer static linking of libraries found by Meson
For all libraries we build in contribs we want the static library version
so our binaries can be distributed.

The `--prefer-static` option tries to use system libraries as static
libraries when both variants are possible [1]. We don't want to have to
distribute the dynamic libraries it may find during contrib builds.

[1] https://mesonbuild.com/Commands.html