Don't bind enqueue pref to one-instance mode
In the simple prefs interface, the option to enqueue files when
using one-instance mode was being disabled if the one-instance-mode
pref is unchecked.

But there's a second pref, one-instance-mode when launched from the
file manager, for which the enqueue option also has meaning. To
ensure that users can always enable enqueue mode, remove its
dependency on the one-instance-mode pref being enabled.

Signed-off-by: FeRD (Frank Dana) <ferdnyc@gmail.com>