vout: make vd->source the new input format on update_format
Internally most display module use vd->source rather than a variable
to use the input format.

If vout_SetDisplayFormat() succeeds the new format must be in vd->source.
On failure vd->source will use the old (chroma) value agreed upon and
we'll still provide pictures
in the format that was agreed upon.

This is the same as when opening a display module. We no longer provide the
source format, the display module relies on vd->source.