vout: allow the display modules to set a video context with the request chroma
If a display modules want to request a GPU chroma when the source is a CPU
chroma, it should be able to do so. In that case the core should use that video
context to provide to filters (for example to handle the CPU to GPU conversion).

Ultimately display modules can just request GPU chromas and do the texture
upload in a converter outside of the display module.

Followup to the discussion on https://code.videolan.org/videolan/vlc/-/merge_requests/3436#note_409707