decoder_helpers: always adjust the dimensions to handle subsampling
There could be some RGB formats that are not 4:4:4. XYZ may also not
be 4:4:4.

Opaque chromas still don't benefit from this adjustment as plane_count is 0.

It will be slightly faster for decoders outputting YUV (95% of them)
as we would have to get the chroma description anyway. But we don't
need the loop in vlc_fourcc_IsYUV() to find it's in there.