input: thumbnailer: remove seek to same position
When tracing the thumbnailing test, we can see that the media was first
opened and demuxed, before DEMUX_SET_POSITION/SET_TIME is sent to the
demux which will reset back to the same initial position.

Check that actual seek values are provided to avoid the seek when not
needed.