input: thumbnailer: ensure time has VLC_TICK_0
vlc_thumbnailer_RequestByTime() is used with a time parameter indicating
the offset from the beginning of the media. Then, the function will
trigger a seek which is expressed in absolute media time, ie.
VLC_TICK_0 + the offset from the beginning of the media.

Callers don't expect to add VLC_TICK_0 themselves, and were not doing
it, but it's needed to ensure that no seek parameter was given.