TODO: es_out: replace by vlc_clock_Start() method
Ensure the clock is now started with vlc_clock_Start() called on the
input vlc_clock_t so that the start point is correctly defined.

Now that vlc_clock_main_Reset() is not called anymore at this location,
the points registered from the input when it is driving the clock bus
will not be removed.

    /!\ As a consequence, --clock-master=input now starts from the input
    /!\ clock points instead of starting from a monotonic clock point.

This change breaks playing files with --clock-master=input.

Co-authored-by: Thomas Guillem <thomas@gllm.fr>