qt: reset playlist current item when goto and play is requested
The interface loads the player view immediately after `addAndPlay()`
is called. This means that the player view will load the content
associated with the old current item.

We can assume that the current item is going to change, because
`vlc_playlist_GoTo()` is used. Thus, it makes sense to reset
the current item meanwhile so that the player view does not
load content just to discard it after the current item is
changed and validated.