clock: compute offset from start_time
Previously, the clock was driven by two values: coeff and offset. The
idea was that since the clock is supposed to look like an affine line,
we could extrapolate any `ts` media point into `ts * coeff + offset` in
system time.

It meant that offset could take any value depending on the difference
between system time and media time, reducing the interpretability of the
values.

After this patch, offset is computed from the start of the stream, so
it will basically only include the difference between the timestamp
computation and the real clock.

Co-authored-by: Thomas Guillem <thomas@gllm.fr>