qml: enable dropping in PlaylistMediaList
With this patch, it is now possible to drag
media library and input items directly
into a playlist.

New input item is created for an unknown
source. So, actions such as dragging from
file browser should also be supported as long
as the drag event contains URLs.

It is also possible to append a playlist into
another playlist.

Previously, it was only possible to first
load the playlist in order to add items.
Now, in the playlist view (the view where
all the playlists are listed) items can be
dropped in to the playlists.